PR 43 (1/4): the epoch ring -- N=3 slots, u64 epochId, deleter release rekeyed to epoch retirement
The boundary package becomes an explicit epoch (refresh doc §1/§2), still
produced at the barrier under the park (consumer unchanged; lockstep):

- SimSnapshot: the front/back double buffer becomes a 3-slot ring. Each slot
  holds all six row namespaces + EpochSlotMeta {u64 epochId (monotonic per
  game), frame span, atomic refCount, per-slot channel-version scalars}.
  Update() is the producer half (extract into a free slot + publish);
  AcquireNewestEpoch() is the consumer half (ref++ newest / ref-- prev),
  called from barrier step 3a; a refcount hitting 0 RETIRES the epoch and
  returns its id for the retirement hooks. Enumerated lockstep exception:
  a no-publish boundary refreshes the net-mutable team/player/global
  channels in place (value-identical to the pre-ring unconditional
  re-extract+swap; dissolves at the 44a producer flip).
- EpochId() (u64) replaces Generation() (u32) as the key of every draw-side
  derived cache (§2.5): cmd-queue cache, piece cache, team-unit index,
  quad-membership mirrors, projectile membership, SnapshotPickGrid, and the
  pregame ==0 fallbacks. Monotonic per game, explicit slots -- the
  ClearCaches cross-game aliasing rationale dissolves.
- DeferredObjectDeleter release rekeyed to epoch retirement (§2.2): the
  split ack (barrier step 8) becomes AckDrainedDestroysEpoch(epochId),
  tagging poisoned shells with the epoch whose record dispatch completed;
  ReleaseRetired(retiredEpochId) returns them to the pools when that epoch
  retires (replaces the end-of-Draw ReleaseAcked + the next-ack-releases-
  leftovers special case). The valve keeps ack+release-all in place (the
  documented lockstep-degenerate immediate retire -- it must free pages
  mid-frame). Flag-off ack/release timing is byte-identical (untouched
  AckDrainedDestroys + end-of-Draw ReleaseAcked).
- Per-slot channel-version scalars (§2.1): cmdQueueCacheEpoch /
  pieceCacheEpoch getters + a DrawMapMirrors::DrainSerial() counter, sealed
  into the held slot at barrier step 3b'' (bookkeeping under lockstep; the
  physical caches track the newest epoch until 44a).
- LogEpochStats(): ring occupancy + per-channel payload bytes (the §7.1
  standing-TODO measurement hook; /epochstats lands with the gate commit).

Flag-off byte-identical: the ring is a pure representation swap (same values
the double buffer held); all split-path changes are SimDrawSplit-gated.

+++ rts/Rendering/Common/SnapshotPickGrid.h
@@ -76,7 +76,7 @@ private:
 	int numX = 0;
 	int numZ = 0;
 	int cellWorld = 0;         // elmos per cell (BASE_QUAD_SIZE)
-	uint32_t builtGeneration = 0;
+	uint64_t builtGeneration = 0; // PR 43 §2.5: EpochId key (u64, monotonic per game)
 	bool built = false;
 
 	// per-cell id lists, indexed [z * numX + x]; capacity is reused across
